pyproject.toml
[project]
name = "photocalendar"
version = "0.1.0"
description = "Calendario fotográfico con Streamlit y LaTeX"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.28",
    "pandas>=2.0",
    "ephem>=4.1",
    "python-dateutil>=2.8",
]
